refactor(move): moving numbers function
