+++ .github/workflows/main.yml
@@ -1,4 +1,4 @@
-on: [push, pull_request]
+on: [pull_request]
 
 jobs:
   main:
@@ -9,6 +9,18 @@ jobs:
         uses: actions/checkout@v2
         with:
           fetch-depth: 0
+      - name: Dump GitHub context
+        env:
+          GITHUB_CONTEXT: ${{ toJson(github) }}
+        run: echo "$GITHUB_CONTEXT"
+      - name: Language
+        env:
+          LANGUAGE: ${{ github.event.repository.language }}
+        run: echo "$LANGUAGE"
       - name: RefDiff
         id: ref-diff
+        env:
+          LANGUAGE: ${{ github.event.repository.language }}
+          REV_BEFORE: ${{ github.event.pull_request.base.sha }}
+          REV_AFTER: ${{ github.event.pull_request.head.sha }}
         uses: rodrigo-brito/refdiff-github-action@master
