+++ MyNumbers.java
@@ -1,0 +1,16 @@
+import java.io.*;
+
+public class MyNumbers {
+	public String Numbers() {
+		String result = "";
+
+		// numbers
+		result += "1";
+		result += "2";
+		result += "3";
+		result += "4";
+		result += "5";
+
+		return result;
+	}
+}
