+++ README.md
@@ -1,2 +1,4 @@
 # refactoring-diff-example
 Toy example for RefDiff Plugin test.
+
+See refactorings in PR opned: https://github.com/rodrigo-brito/refactoring-diff-example/pull/3/files
